// ui/lib/api/schemas/fontFaceInfo.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { FontSource } from './fontSource'

export interface FontFaceInfo {
  cached: boolean
  /** @nullable */
  category?: string | null
  familyName: string
  postScriptName: string
  source: FontSource
}