// ui/lib/api/schemas/nodeDataPatch.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { ImageDataPatch } from './imageDataPatch'
import type { MaskDataPatch } from './maskDataPatch'
import type { TextDataPatch } from './textDataPatch'

export type NodeDataPatch =
  | {
      text: TextDataPatch
    }
  | {
      image: ImageDataPatch
    }
  | {
      mask: MaskDataPatch
    }